+++ StudyDesck.API/Domain/Models/SessionMaterial.cs
@@ -10,7 +10,7 @@ namespace StudyDesck.API.Domain.Models
         public Session Session { get; set; }
         public int SessionId { get; set; }
         public StudyMaterial StudyMaterial { get; set; }
-        public long StudyMaterialId { get; set; }
+        public int StudyMaterialId { get; set; }
         public Tutor Tutor { get; set; }
         public int TutorId { get; set; }
     }
