+++ StudyDesck.API/Domain/Models/StudyMaterial.cs
@@ -7,7 +7,7 @@ namespace StudyDesck.API.Domain.Models
 {
     public class StudyMaterial
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
         public string FileUrl { get; set; }
