+++ StudyDesck.API/Domain/Persistence/Repositories/ISessionMaterialRepository.cs
@@ -11,12 +11,12 @@ namespace StudyDesck.API.Domain.Persistence.Repositories
         Task<IEnumerable<SessionMaterial>> ListBySessionIdAsync(int sessionId);
         Task<IEnumerable<SessionMaterial>> ListByStudyMaterialIdAsync(int studyMaterialId);
         Task<IEnumerable<SessionMaterial>> ListByTutorIdAsync(int tutorId);
-        Task<SessionMaterial> FindBySessionIdAndStudyMaterial(int sessionId, long studyMaterial);
+        Task<SessionMaterial> FindBySessionIdAndStudyMaterial(int sessionId, int studyMaterial);
         Task AddAsync(SessionMaterial sessionMaterial);
         void Remove(SessionMaterial sessionMaterial);
 
-        Task AssignSessionMaterial(int sessionId, long materialId, int tutorId);
-        Task AssignSessionMaterial(int sessionId, long materialId);
-        Task UnassignStudyMaterial(int sessionId, long materialId);
+        Task AssignSessionMaterial(int sessionId, int materialId, int tutorId);
+        Task AssignSessionMaterial(int sessionId, int materialId);
+        Task UnassignStudyMaterial(int sessionId, int materialId);
     }
 }
