+++ StudyDesck.API/Domain/Persistence/Repositories/IStudentMaterialRepository.cs
@@ -11,12 +11,12 @@ namespace StudyDesck.API.Domain.Persistence.Repositories
         Task<IEnumerable<StudentMaterial>> ListByStudentIdAsync(int studentId);
         Task<IEnumerable<StudentMaterial>> ListByCategoryIdAsync(int categoryId);
         Task<IEnumerable<StudentMaterial>> ListByInstituteIdAsync(int instituteId);
-        Task<StudentMaterial> FindByStudentIdAndStudyMaterialId(int studentId, long studyMaterial);
+        Task<StudentMaterial> FindByStudentIdAndStudyMaterialId(int studentId, int studyMaterial);
         Task AddAsync(StudentMaterial studentMaterial);
         void Remove(StudentMaterial studentMaterial);
 
-        Task AssignStudentMaterial(int studentId, long materialId, int categoryId, int instituteId); // TODO to see
-        Task AssignStudentMaterial(int studentId, long materialId); // TODO to see
-        Task UnassignstudyMaterial(int studentId, long materialId); // TODO to see
+        Task AssignStudentMaterial(int studentId, int materialId, int categoryId, int instituteId); // TODO to see
+        Task AssignStudentMaterial(int studentId, int materialId); // TODO to see
+        Task UnassignstudyMaterial(int studentId, int materialId); // TODO to see
     }
 }
