+++ StudyDesck.API/Domain/Services/IStudentMaterialService.cs
@@ -10,8 +10,8 @@ namespace StudyDesck.API.Domain.Services
     public interface IStudentMaterialService
     {
         Task<IEnumerable<StudentMaterial>> ListByStudentIdAsync(int studentId);
-        Task<StudentMaterialResponse> AssignStudentMaterialAsync(int studentId, long materialId);
-        Task<StudentMaterialResponse> UnassignStudentMaterialAsync(int studentId, long materialId);
+        Task<StudentMaterialResponse> AssignStudentMaterialAsync(int studentId, int materialId);
+        Task<StudentMaterialResponse> UnassignStudentMaterialAsync(int studentId, int materialId);
         Task<StudentMaterialResponse> AssignStudentMaterialAsync(int studentId, StudentMaterial studentMaterial);
     }
 }
